docs:working in new main map example (#128) * update dart sdk
