+++ example/lib/src/search_example.dart
