+++ example/lib/src/simple_example.dart
