+++ pubspec.yaml
@@ -4,7 +4,7 @@ version: 0.15.0
 
 homepage: https://github.com/liodali/osm_flutter
 environment:
-  sdk: '>=2.12.0 <3.0.0'
+  sdk: '>=2.14.0 <3.0.0'
   flutter: ">=1.12.0"
 dependencies:
   flutter:
